+++ crates/actions/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod actions;
+pub mod operations;
 mod utils;